requirements.txt
environ==1.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-environ==0.4.5
django-cors-headers==3.1.1
django-rest-swagger==2.2.0
django-summernote==0.8.11.4
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jsonapi==3.0.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.4
psycopg2==2.7
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.6
python-dotenv==0.10.3